src/test/java/jdbc/jdbc/repository/MemberRepositoryV0Test.java
package jdbc.jdbc.repository;

import static org.assertj.core.api.Assertions.*;
import static org.junit.jupiter.api.Assertions.*;

import java.util.NoSuchElementException;
import jdbc.jdbc.domain.Member;
import lombok.extern.slf4j.Slf4j;
import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;
import org.slf4j.LoggerFactory;

@Slf4j
class MemberRepositoryV0Test {
    MemberRepositoryV0 repository = new MemberRepositoryV0();
    @Test
    public void create() throws Exception{
    //given
        Member member = new Member("memberV8", 10000);
    //when
        repository.save(member);
    //then
        Member findMember = repository.findById(member.getMemberId());
        log.info("findMember={}", findMember);
        assertThat(findMember).isEqualTo(member);

        repository.update(member.getMemberId(), 20000);
        assertThat(repository.findById(member.getMemberId()).getMoney()).isEqualTo(20000);
        repository.delete(member.getMemberId());
        assertThatThrownBy(() -> repository.findById(member.getMemberId())).isInstanceOf(
            NoSuchElementException.class);

    }
}